requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
eventlet==0.33.3
requests==2.31.0
SQLAlchemy==2.0.23
python-dotenv==1.0.1
gunicorn==21.2.0
Pillow==10.2.0
numpy==1.26.4
